Split configuration from main file.
